Day 09: Optimize part 2.
We used maths behind it.
We "draw" a line to the right (constant y) and change state (inside or outside) each time we cross the polygon's boundary.
We ensure that the point lies to the left (our left) of the boundary before changing state, hence the "less than" operation.

The operation doesn't change whether the boundary points up or down. We always need to use the "less than" operation.